refactored bpm-form component to use bulma txt fields and buttons. Added styling to form and button components
