+++ package.json
@@ -6,10 +6,14 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "bulma": "^0.9.3",
+    "bulma-styled-components": "^1.0.2",
+    "node-sass": "^4.14.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-player": "^2.9.0",
     "react-scripts": "4.0.3",
+    "sass-loader": "^12.1.0",
     "use-sound": "^4.0.1",
     "web-vitals": "^1.0.1"
   },
