+++ src/App.scss
@@ -1,0 +1,2 @@
+@import 'bulma/bulma.sass';
+
