+++ src/components/bpm-form/bpm-form-styles.scss
@@ -1,0 +1,1 @@
+@import 'bulma/bulma.sass';
