+++ src/components/fart-button/fart-button-styles.scss
@@ -1,0 +1,2 @@
+@import 'bulma/bulma.sass';
+
